Fix some Cppcheck warnings.

+++ src/Styles.c
@@ -3515,7 +3515,7 @@ bool Style_SelectColor(HWND hwnd, LPWSTR lpszStyle, int cchStyle, bool bFore) {
 		if (StrNotEmpty(szNewStyle)) {
 			lstrcat(szNewStyle, L"; ");
 		}
-		wsprintf(tch, L"fore:#%06X", iRGBResult);
+		wsprintf(tch, L"fore:#%06X", (unsigned)iRGBResult);
 		lstrcat(szNewStyle, tch);
 		Style_StrCopyBack(szNewStyle, lpszStyle, tch);
 	} else {
@@ -3523,7 +3523,7 @@ bool Style_SelectColor(HWND hwnd, LPWSTR lpszStyle, int cchStyle, bool bFore) {
 		if (StrNotEmpty(szNewStyle)) {
 			lstrcat(szNewStyle, L"; ");
 		}
-		wsprintf(tch, L"back:#%06X", iRGBResult);
+		wsprintf(tch, L"back:#%06X", (unsigned)iRGBResult);
 		lstrcat(szNewStyle, tch);
 	}
